Fix reporting of "share" button (KEY_RECORD)
Tested on a 8BitPro Pro 2 Wired Controller for Xbox. The driver was not
reporting button presses of the "share" button (below the big Xbox
button). After confirming that the controller is detected as "Series XS"
and analyzing packet dumps, I noticed that this controller reports the
button in a different offset than this driver expects.

Since the controller works in Windows with the stock, official Microsoft
driver I am assuming that both offsets are being used in the wild to
report the state of this button.

Signed-off-by: Leonardo Brondani Schenkel <leonardo@schenkel.net>