Implement GIP authentication
The authentication handshake uses either an ECDH or RSA key exchange,
depending on the device that is used. In the case of RSA,
an X.509 certificate has to be parsed during the handshake.
Unfortunately, the certificates issued by Microsoft do not comply
with RFC 5280. For this reason, the `x509_cert_parse` function cannot
be used to parse the certificates. Instead, the pubkey is extracted
from the certificates by finding the start of the ASN.1 sequence.